add_desat

Add the desaturation stops to the dive curve. It also check if there is no depth point higher and before the stops, to prevent accidents.

add_desat(dtcurve, desat, ascent_speed = 10, secu = FALSE)

Arguments

dtcurve
a depth time curve in a data.frame with 2 columns depth and time. Depths values are meters (positive values) and time is in minute.

desat
a desat object that contain the desaturation stops of given dive

ascent_speed
Ascent_speed in meter/minute. 10 m/min by default. Most dive table advice to limite this speed to 20M/min maximum.

secu
security decompression stop of 3 min at 3 m. FALSE by default.

Value

a dtcurve data.frame with the same format, but desaturation stop have been rbinded at the end.

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>