dive

Create a dive class object simulating a dive for given depth and times. To do so, it create a depth/time curve and find the desaturation stops according to a desaturation model.

dive(
  depth = 20,
  time = 40,
  secu = TRUE,
  ascent_speed = 10,
  maj = 0,
  desat_model = c("table", "other"),
  hour = 0,
  way = c("OW", "WB")
)

Arguments

depth
Depth of the dive in meter. Need to be positive values. A single value is needed for square dive, however if a vector is provided a decompression model need to be selected. Default is 20 meter

time
Duration of the dive in minute. Need to be positive values. A single value is needed for square dive, however if a vector is provided a decompression model need to be selected. Default is 40 minutes

secu
security decompression stop of 3 min at 3 m. FALSE by default.

ascent_speed
Ascent_speed in meter/minute. 10 m/min by default. Most dive table advice to limite this speed to 20M/min maximum.

maj
Time majoration for the dive. Only used by table decompression model.

desat_model
Which desaturation model is used to compute desaturation stops during ascent, to eliminate nitrogen. Default is tables as only this one works today.

hour
The first immersion hour in minute. Need to be 24h format converted in minutes (hour = hours * 60 + minutes). 0 by default.

way
If the dive is one way (one way : 'OW') or if the diver return by the same depth (way back : 'WB').

Value

dive, a list containing a depth/time curve in a data.frame, the desaturation stops acording to the model used, and some information about depth to surface time, and input parameters as ascent_speed and secu.

Details

See tablecheck for limit values of depth and time.

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Examples

dive = dive(depth = 39, time = 22, secu = TRUE, ascent_speed = 10)