dtcurve

Trace a curve of depth and time for given parameters and palier information

dtcurve(
  time,
  depth,
  palier,
  vup = 10,
  dist = NULL,
  speed = NULL,
  way = c("AR", "AS")
)

Arguments

time
in minute. Max is 180' see Detail !

depth
in meter. Max is 65m see Detail !

palier
a object palier computed with the palier function. Is a list of depth and time for every deco stage.

vup
10 m/min max speed up to the deco. Speed between deco is fixed to 6

dist
a distance vector

speed
speed of the diver

way
if the dive is one way or the diver return by the same depth

Value

a list, with a vector of depths, a vector of time and the dtr value.

Details

This function will stop if the depth > 65 or time > 180 because the table are limited to this extent. However for lower values the table can return NA values.

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>