plot.dive

Plot the dive curve depending on time and depth. Only represent square profile

# S3 method for dive
plot(
  x,
  ...,
  dtr_print = FALSE,
  hour_print = TRUE,
  line_print = TRUE,
  deco_print = TRUE,
  depth_print = TRUE,
  time_print = TRUE,
  def_cols = FALSE,
  legend = FALSE,
  add = FALSE
)

Arguments

x
an object of class dive.

...
every argument for the plot function such as graphical parameters for lines. Classical graphical parameter applies on the dive curve.

type for the size line, set to 'b' by default.

xlab set to 'Time' by default, see title.

ylab set to 'Depth' by default, see title.

dtr_print
set to FALSE by default, whether there is text and line for ascent time.

hour_print
set to TRUE by default, whether there is the hour at beginning and end of the dive.

line_print
set to TRUE by default, whether there is lines for depths.

deco_print
set to TRUE by default, whether there is information about desat stop for depths and times.

depth_print
set to TRUE by default, whether there is depth values along the dive.

time_print
set to TRUE by default, whether there is time values along the dive.

def_cols
set to FALSE by default, modify plot colors to match shinyapp css.

legend
set to FALSE by default, print legend about the ascent speeds.

add
set to FALSE by default, to add another dive plot on a precedent one.

See also

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>