rules

rules retrieve the rules of a tank or pressure and times at which the pressure of tank(s) met the rules in conso.

rules(object, ...)

# S3 method for tank
rules(object, ...)

# S3 method for conso
rules(object, ..., hour = FALSE)

Arguments

object
is a DiveR object. There is methods for tank and conso objects.

...
other parameters passed to the pressure.conso method.

tankn

the number of the tank if there is multiple. NULL by default will return all the tanks

n

the number of the rule. 0 mean the air-failure, 1 rule1 and 2 rule2. NULL by default will get all of them. A vector of values can be provided

hour
if the time defined inside the hours of the dive (TRUE) or in relative time to dive start (FALSE). Set to FALSE by default.

Value

for tank

numeric vector for rules in their order

for conso

data.frame with rules in column and tanks in rows

@examples t <- tank(12,200) rules(object = t)

d <- dive(20,15) c <- conso(d, t, cons = 20) rules(object = c)

rules(object = c, n = 0)

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>