summary

summary of a dive object

# S3 method for dive
summary(object, ...)

# S3 method for palier
summary(object, ...)

# S3 method for tank
summary(object, ...)

# S3 method for conso
summary(object, ...)

Arguments

object
is a DiveR object. There are methods for dive, tank, conso and palier objects.

...
other arguments passed to the dtr.palier method.

plot

FALSE by default

Value

A brief summary of a dive and its main parameters.

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Examples

summary(palier(depth = 20, time = 50))

#> 1 stops :   7 minutes at 3 meters 
#> The dive group is I

summary(palier(depth = 20, time = 40,secu = FALSE))

#> 0 stops : no stop for deco.
#>   Please consider a minimal safety 
#>                  stop of 3 minute at 3 meter 
#> The dive group is H

summary(tank(vol = 12, press = 200))

#> Tank : 12 litre at 200 bar
#> rules : 1 mid : 100 bar
#>         2 res :  50 bar
#> The tank type is back and contain Air 
#> Named : back12 

Tank_15L <- tank(vol = 15, press = 200)
viable <- conso(dive = dive(20,40), tank = Tank_15L, 
                cons = 20, failure_label = 'Air failure')
summary(viable)

#> Consumption simulated on dive at 20 m for 45.2 minutes
#> ---------------------------------------------------------------------
#>        Tank name |         Rule | Pressure |    Time | Final pressure 
#> ---------------------------------------------------------------------
#>      Tank back15 |          mid |  100 bar |  25 min | 29.16 bar
#>                  |          res |   50 bar |  38 min |    
#>                  |  Air failure |    0 bar |  NA min |     
#> ---------------------------------------------------------------------
#> The dive is viable !