summary

summary method for class dive.

# S3 method for tank
summary(object, ...)

# S3 method for conso
summary(object, ...)

# S3 method for dive
summary(object, ...)

# S3 method for desat
summary(object, ...)

Arguments

object
is a DiveR object. There are methods for dive, tank and conso objects.

...
other arguments not used

Value

A brief summary of a dive and its main parameters.

summary(dive(20, 40, secu = FALSE)) summary(dive(20, 40, secu = TRUE)) summary(dive(20, 45)) summary(dive(39, 22)) summary(dive(50, 22))

Author

Jaunatre Maxime <maxime.jaunatre@yahoo.fr>

Examples

summary(tank(vol = 12, press = 200))

#> Tank : 12 litre at 200 bar
#> rules : 1 mid : 100 bar
#>         2 res :  50 bar
#> The tank type is back and contain Air 
#> Named : back12 

Tank_15L <- tank(vol = 15, press = 200)
viable <- conso(dive = dive(20,40), tank = Tank_15L, 
                cons = 20, failure_label = 'Air failure')
summary(viable)

#> Consumption simulated on dive at 20 m for 45.2 minutes
#> ---------------------------------------------------------------------
#>        Tank name |         Rule | Pressure |    Time | Final pressure 
#> ---------------------------------------------------------------------
#>      Tank back15 |          mid |  100 bar |  25 min | 29.16 bar
#>                  |          res |   50 bar |  38 min |    
#>                  |  Air failure |    0 bar |  NA min |     
#> ---------------------------------------------------------------------
#> The dive is viable !

summary(desat_table(init_dtcurve(20, 40)))

#> --------- No desat stop ---------
#>     Group : H | Model :   table 

summary(desat_table(init_dtcurve(20, 45)))

#> ---------------------------------
#>  Stop | Depth | Duration |   Time 
#> ---------------------------------
#>  n  3 |   3 m |    1 min | NA min 
#> ---------------------------------
#>     Group : I | Model :   table 

summary(desat_table(init_dtcurve(39, 22)))

#> ---------------------------------
#>  Stop | Depth | Duration |   Time 
#> ---------------------------------
#>  n  2 |   6 m |    2 min | NA min 
#> ---------------------------------
#>  n  3 |   3 m |   19 min | NA min 
#> ---------------------------------
#>     Group : J | Model :   table 

summary(desat_table(init_dtcurve(50, 22)))

#> ---------------------------------
#>  Stop | Depth | Duration |   Time 
#> ---------------------------------
#>  n  1 |   9 m |    1 min | NA min 
#> ---------------------------------
#>  n  2 |   6 m |    8 min | NA min 
#> ---------------------------------
#>  n  3 |   3 m |   32 min | NA min 
#> ---------------------------------
#>     Group : L | Model :   table